create sanity - break out monolithic js files into something more workable
