+++ package.json
@@ -2,7 +2,7 @@
   "name": "discordcsgoserverbot",
   "version": "7.0.0",
   "description": "Server info bot",
-  "main": "index.js",
+  "main": "src/index.js",
   "type": "module",
   "scripts": {
     "start": "node index.js",
